% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmspc-package.R
\docType{package}
\name{rmspc-package}
\alias{rmspc}
\alias{rmspc-package}
\title{rmspc: Multiple Sample Peak Calling}
\description{
The rmspc package runs MSPC (Multiple Sample 
    Peak Calling) software using R. The analysis of ChIP-seq 
    samples outputs a number of enriched regions (commonly 
    known as "peaks"), each indicating a protein-DNA interaction 
    or a specific chromatin modification. When replicate samples 
    are analyzed, overlapping peaks are expected. This repeated 
    evidence can therefore be used to locally lower the minimum 
    significance required to accept a peak. MSPC uses combined evidence 
    from replicated experiments to evaluate peak calling output, 
    rescuing peaks, and reduce false positives. It takes any 
    number of replicates as input and improves sensitivity and 
    specificity of peak calling on each, and identifies consensus 
    regions between the input samples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://genometric.github.io/MSPC/}
  \item Report bugs at \url{https://github.com/Genometric/MSPC/issues}
}

}
\author{
\strong{Maintainer}: Meriem Bahda \email{meriembahda@gmail.com}

Authors:
\itemize{
  \item Vahid Jalili \email{jalili.vahid@gmail.com}
  \item Marzia Angela Cremona \email{marzia.cremona@fsa.ulaval.ca}
  \item Fernando Palluzzi \email{fernando.palluzzi@gmail.com}
}

}
\keyword{internal}
