% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmspc-package.R
\docType{package}
\name{rmspc-package}
\alias{rmspc}
\alias{rmspc-package}
\title{rmspc: MSPC program}
\description{
The rmspc package runs the MSPC software using R. 
    The analysis of ChIP-seq samples outputs a number of 
    enriched regions (commonly known as "peaks"), each
    indicating a protein-DNA interaction or a specific
    chromatin modification. When replicate samples are
    analyzed, overlapping peaks are expected. This repeated
    evidence can therefore be used to locally lower the minimum
    significance required to accept a peak. Here, we propose a
    method for joint analysis of weak peaks. Given a set of 
    peaks from (biological or technical) replicates, the method
    combines the p-values of overlapping enriched regions, and
    allows the "rescue" of weak peaks occurring in more than 
    one replicate.
    MSPC comparatively evaluates ChIP-seq peaks and combines
    the statistical significance of repeated evidences, 
    with the aim of lowering the minimum significance required
    to rescue weak peaks; hence reducing false-negatives.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://genometric.github.io/MSPC/}
  \item Report bugs at \url{https://github.com/Genometric/MSPC/issues}
}

}
\author{
\strong{Maintainer}: Meriem Bahda \email{meriembahda@gmail.com}

Authors:
\itemize{
  \item Vahid Jalili \email{jalili.vahid@gmail.com}
}

Other contributors:
\itemize{
  \item Marzia Angela Cremona \email{marzia.cremona@gmail.com} [contributor]
}

}
\keyword{internal}
